Allow skipping gsoap build in boostrap Allow defining custom gsoap src location

+++ bootstrap.sh
@@ -1,20 +1,31 @@
+SKIP_GSOAP=0
+i=1;
+for arg in "$@" 
+do
+    if [ $arg == "--skip-gsoap" ]; then
+        SKIP_GSOAP=1
+    fi
+    i=$((i + 1));
+done
+
 sudo apt install automake autoconf gcc make pkg-config
 sudo apt install unzip
 
-echo "-- Building gsoap libgsoap-dev --"
-#WS-Security depends on OpenSSL library 3.0 or 1.1
-wget https://sourceforge.net/projects/gsoap2/files/gsoap_2.8.123.zip/download
-unzip download
-rm download
-cd gsoap-2.8
-mkdir build
-./configure --with-openssl=/usr/lib/ssl --prefix=$(cd $(dirname "${BASH_SOURCE[0]}") && pwd)/build
-LIBRARY_PATH="$(pkg-config --variable=libdir openssl):$LIBRARY_PATH" \
-LD_LIBRARY_PATH="$(pkg-config --variable=libdir openssl):$LD_LIBRARY_PATH" \
-    make -j$(nproc)
-make install
-cd ..
-
+if [ $SKIP_GSOAP -eq 0 ]; then
+    echo "-- Building gsoap libgsoap-dev --"
+    #WS-Security depends on OpenSSL library 3.0 or 1.1
+    wget https://sourceforge.net/projects/gsoap2/files/gsoap_2.8.123.zip/download
+    unzip download
+    rm download
+    cd gsoap-2.8
+    mkdir build
+    ./configure --with-openssl=/usr/lib/ssl --prefix=$(cd $(dirname "${BASH_SOURCE[0]}") && pwd)/build
+    LIBRARY_PATH="$(pkg-config --variable=libdir openssl):$LIBRARY_PATH" \
+    LD_LIBRARY_PATH="$(pkg-config --variable=libdir openssl):$LD_LIBRARY_PATH" \
+        make -j$(nproc)
+    make install
+    cd ..
+fi
 
 aclocal
 autoconf
